Modify checklist intermediate service
