+++ checklist-dashboard/widget/widget-backend/mpr-backend-service.iml
@@ -1,0 +1,74 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module org.jetbrains.idea.maven.project.MavenProjectsManager.isMavenModule="true" type="JAVA_MODULE" version="4">
+  <component name="NewModuleRootManager" LANGUAGE_LEVEL="JDK_1_8">
+    <output url="file://$MODULE_DIR$/target/classes" />
+    <output-test url="file://$MODULE_DIR$/target/test-classes" />
+    <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/src/main/java" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/src/main/resources" type="java-resource" />
+      <excludeFolder url="file://$MODULE_DIR$/target" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+    <orderEntry type="library" name="Maven: org.wso2.msf4j:msf4j-core:2.5.2" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.carbon.messaging:org.wso2.carbon.messaging:3.0.2" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.carbon:org.wso2.carbon.launcher:5.2.0" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.transport.http:org.wso2.transport.http.netty:6.0.64" level="project" />
+    <orderEntry type="library" name="Maven: io.netty:netty-common:4.1.16.Final" level="project" />
+    <orderEntry type="library" name="Maven: io.netty:netty-buffer:4.1.16.Final" level="project" />
+    <orderEntry type="library" name="Maven: io.netty:netty-transport:4.1.16.Final" level="project" />
+    <orderEntry type="library" name="Maven: io.netty:netty-handler:4.1.16.Final" level="project" />
+    <orderEntry type="library" name="Maven: io.netty:netty-codec:4.1.16.Final" level="project" />
+    <orderEntry type="library" name="Maven: io.netty:netty-handler-proxy:4.1.16.Final" level="project" />
+    <orderEntry type="library" name="Maven: io.netty:netty-codec-socks:4.1.16.Final" level="project" />
+    <orderEntry type="library" name="Maven: io.netty:netty-codec-http2:4.1.16.Final" level="project" />
+    <orderEntry type="library" name="Maven: io.netty:netty-resolver:4.1.16.Final" level="project" />
+    <orderEntry type="library" name="Maven: commons-pool.wso2:commons-pool:1.5.6.wso2v1" level="project" />
+    <orderEntry type="library" name="Maven: commons-pool:commons-pool:1.5.6" level="project" />
+    <orderEntry type="library" name="Maven: javax.websocket:javax.websocket-api:1.1" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.eclipse.osgi:org.eclipse.osgi:3.10.2.v20150203-1939" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.eclipse.osgi:org.eclipse.osgi.services:3.4.0.v20140312-2051" level="project" />
+    <orderEntry type="library" name="Maven: javax.ws.rs:javax.ws.rs-api:2.0" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.msf4j:jaxrs-delegates:2.5.2" level="project" />
+    <orderEntry type="library" name="Maven: org.apache.httpcomponents:httpclient:4.5.2" level="project" />
+    <orderEntry type="library" name="Maven: org.apache.httpcomponents:httpcore:4.4.4" level="project" />
+    <orderEntry type="library" name="Maven: commons-codec:commons-codec:1.9" level="project" />
+    <orderEntry type="library" name="Maven: org.slf4j:slf4j-api:1.7.5" level="project" />
+    <orderEntry type="library" name="Maven: org.slf4j:slf4j-log4j12:1.6.0" level="project" />
+    <orderEntry type="library" name="Maven: log4j:log4j:1.2.14" level="project" />
+    <orderEntry type="library" name="Maven: com.google.code.gson:gson:2.2.4" level="project" />
+    <orderEntry type="library" name="Maven: com.google.code.findbugs:jsr305:2.0.1" level="project" />
+    <orderEntry type="library" name="Maven: org.apache.servicemix.bundles:org.apache.servicemix.bundles.commons-beanutils:1.8.3_2" level="project" />
+    <orderEntry type="library" name="Maven: commons-logging:commons-logging:1.1.1" level="project" />
+    <orderEntry type="library" name="Maven: com.nimbusds:nimbus-jose-jwt:2.25" level="project" />
+    <orderEntry type="library" name="Maven: net.jcip:jcip-annotations:1.0" level="project" />
+    <orderEntry type="library" name="Maven: net.minidev:json-smart:1.1.1" level="project" />
+    <orderEntry type="library" name="Maven: org.bouncycastle:bcprov-jdk15on:1.50" level="project" />
+    <orderEntry type="library" name="Maven: org.yaml:snakeyaml:1.17" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.carbon.config:org.wso2.carbon.config:2.1.5" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.carbon.secvault:org.wso2.carbon.secvault:5.0.8" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.carbon.utils:org.wso2.carbon.utils:2.0.2" level="project" />
+    <orderEntry type="library" name="Maven: commons-io.wso2:commons-io:2.4.0.wso2v1" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.carbon.dashboards:org.wso2.carbon.dashboards.core:4.0.15" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.carbon.datasources:org.wso2.carbon.datasource.core:1.1.2" level="project" />
+    <orderEntry type="library" name="Maven: com.zaxxer:HikariCP:2.4.1" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.carbon.jndi:org.wso2.carbon.jndi:1.0.3" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.carbon.analytics-common:org.wso2.carbon.analytics.idp.client:6.0.51" level="project" />
+    <orderEntry type="library" name="Maven: io.github.openfeign:feign-core:9.5.0" level="project" />
+    <orderEntry type="library" name="Maven: org.jvnet:animal-sniffer-annotation:1.0" level="project" />
+    <orderEntry type="library" name="Maven: io.github.openfeign:feign-gson:9.5.0" level="project" />
+    <orderEntry type="library" name="Maven: com.googlecode.json-simple.wso2:json-simple:1.1.wso2v1" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.carbon.analytics-common:org.wso2.carbon.analytics.permissions:6.0.51" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.carbon.analytics-common:org.wso2.carbon.database.query.manager:6.0.51" level="project" />
+    <orderEntry type="library" name="Maven: org.osgi:org.osgi.core:6.0.0" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.carbon.uiserver:org.wso2.carbon.uiserver:0.19.5" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.carbon:org.wso2.carbon.core:5.2.5" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.carbon.deployment:org.wso2.carbon.deployment.engine:5.2.0" level="project" />
+    <orderEntry type="library" name="Maven: io.netty:netty-codec-http:4.0.30.Final" level="project" />
+    <orderEntry type="library" name="Maven: com.google.guava:guava:21.0" level="project" />
+    <orderEntry type="library" name="Maven: org.apache.commons:commons-lang3:3.1" level="project" />
+    <orderEntry type="library" name="Maven: org.wso2.orbit.com.github.jknack:handlebars:4.0.3.wso2v1" level="project" />
+    <orderEntry type="library" name="Maven: org.antlr:antlr4-runtime:4.5.1-1" level="project" />
+    <orderEntry type="library" name="Maven: org.apache.felix:org.osgi.core:1.0.0" level="project" />
+  </component>
+</module>
