--- checklist-dashboard/widget/widget-backend/src/main/java/org/wso2/checklist/utils/DataValueHolder.java
@@ -1,43 +1,0 @@
-/*
- * Copyright (c) 2019, WSO2 Inc. (http://wso2.com) All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.wso2.checklist.utils;
-
-import javax.sql.DataSource;
-
-/**
- * This class holding Data source object.
- */
-public class DataValueHolder {
-    private DataSource datasource;
-    private static DataValueHolder dataValueHolder =  new DataValueHolder();
-
-    private DataValueHolder() {
-
-    }
-
-    public static DataValueHolder getInstance() {
-        return dataValueHolder;
-    }
-
-    public void setDataSource(DataSource datasource) {
-        this.datasource = datasource;
-    }
-
-    public DataSource getDataSource() {
-        return datasource;
-    }
-}
